接口接入优化测试

+++ app/src/main/java/com/yjkj/service_recoder/java/dataBaseDao/UserDao.java
@@ -28,7 +28,7 @@ public interface UserDao {
     List<UserBean> getAllUser();
 
     @Query("SELECT * FROM UserBean where userId = :userId")
-    UserBean getUserById(int userId);
+    UserBean getUserById(Long userId);
 
     @Update
     void updateUser(UserBean userBean);
